pyproject.toml
[project]
name = "ler-sheets-em-paralelo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=38.2.0",
    "git-filter-repo>=2.47.0",
    "google-api-python-client>=2.187.0",
    "google-auth>=2.43.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.2",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]
